Time: 0 ms (100.00%) | Memory: 65.1 MB (66.41%) - LeetSync

+++ 713-subarray-product-less-than-k/subarray-product-less-than-k.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int numSubarrayProductLessThanK(vector<int>& nums, int k) {
         int res = 0, l = 0;
-        long product = 1;
+        int product = 1;
 
         for (int r = 0; r < nums.size(); ++r) {
             product *= nums[r];
